refacto(avatars): default avatar and user scoped avatars

+++ src/users/user-avatars/user-avatars.service.ts
@@ -1,11 +1,14 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, Logger } from '@nestjs/common';
 import { FileUploadService } from '../../services/file-upload/file-upload.service';
 import { ConfigService } from '@nestjs/config';
 
 @Injectable()
 export class UserAvatarsService {
+  private logger = new Logger(UserAvatarsService.name);
   private readonly avatarBucket: string;
-  private readonly avatarPrefix: string = 'avatars/';
+  private readonly avatarPrefix: string = 'avatars';
+
+  public static readonly DEFAULT_AVATAR: string = 'avatar.png';
 
   constructor(
     private fileUploadService: FileUploadService,
@@ -14,8 +17,12 @@ export class UserAvatarsService {
     this.avatarBucket = this.configService.getOrThrow('S3_publicBucket');
   }
 
-  async create(userId: string, avatar: Express.Multer.File) {
-    const fileName = `${this.avatarPrefix}${userId}`;
+  async create(
+    userId: string,
+    avatarName: string,
+    avatar: Express.Multer.File,
+  ) {
+    const fileName = `${this.avatarPrefix}/${userId}/${avatarName}`;
 
     await this.fileUploadService.uploadFile(
       this.avatarBucket,
@@ -23,16 +30,27 @@ export class UserAvatarsService {
       avatar,
     );
 
-    return 'Avatar uploaded';
+    return avatarName;
   }
 
-  async findOne(userId: string) {
-    const fileName = `${this.avatarPrefix}${userId}`;
+  async findOne(userId: string, avatarName: string) {
+    let fileName = `${this.avatarPrefix}/${userId}/${avatarName}`;
+
+    // Default avatar not scoped to user
+    if (avatarName === UserAvatarsService.DEFAULT_AVATAR) {
+      fileName = `${this.avatarPrefix}/${avatarName}`;
+    }
+
     return this.fileUploadService.getFile(this.avatarBucket, fileName);
   }
 
-  remove(userId: string) {
-    const fileName = `${this.avatarPrefix}${userId}`;
+  remove(userId: string, avatarName: string) {
+    if (avatarName === UserAvatarsService.DEFAULT_AVATAR) {
+      this.logger.warn('Cannot remove default avatar');
+      return;
+    }
+
+    const fileName = `${this.avatarPrefix}/${userId}/${avatarName}`;
     return this.fileUploadService.removeFile(this.avatarBucket, fileName);
   }
 }
